requirements.txt
Flask
Flask-Cors
requests
pandas
entsoe-pandas
python-dotenv
flake8
